New performance fix, now it checks if there is a need to do delete before performing it, it needs extra select but that is less expensive than delete (croned it's still faster) this needs to be reviewed by someone who knows sql, it's possible that executing empty delete (0 rows are matched by condition) doesn't take more time than than executing empty select + it's even possible that it doesn't replicate db in that case if that is true, fix me + some fixes :)
git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@103146 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ OnlineStatusBar.status.php
@@ -38,7 +38,7 @@ class OnlineStatusBar_StatusCheck {
 			if ( $user->isLoggedIn() ) {
 				$status = $user->getOption( 'OnlineStatusBar_status', $wgOnlineStatusBarDefaultOnline );
 				if ( $delayed_check ) {
-					if ( $result < $w_time ) {
+					if ( $result < wfTimestamp( TS_MW, $w_time ) ) {
 						$status = 'write';
 					}
 				}
@@ -131,6 +131,14 @@ class OnlineStatusBar_StatusCheck {
 			return 0;
 		}
 		$dbw = wfGetDB( DB_MASTER );
+		$t_time = OnlineStatusBar::getTimeoutDate();
+		$result = $dbw->selectField( 'online_status', 'timestamp', array( "timestamp < " . $dbw->addQuotes( $dbw->timestamp( $t_time ) ) ),
+			__METHOD__, array( 'LIMIT 1' ) );
+		if ( $result === false ) {
+			// no need for delete
+			return 0;
+		}
+
 		// calculate time and convert it back to mediawiki format
 		$time = OnlineStatusBar::getTimeoutDate();
 		$dbw->delete( 'online_status', array( "timestamp < " . $dbw->addQuotes( $dbw->timestamp( $time ) ) ), __METHOD__ );
